--- main/mrbc_esp32_tof.h
@@ -1,8 +1,0 @@
-#ifndef MRBC_ESP32_TOF_H
-#define MRBC_ESP32_TOF_H
-
-#include "mrubyc.h"
-
-void mrbc_mruby_esp32_tof_gem_init(struct VM*);
-
-#endif // MRBC_ESP32_TOF_H
