+++ main/mrbc_esp32_vl53l0x.h
@@ -1,0 +1,8 @@
+#ifndef MRBC_ESP32_VL53L0X_H
+#define MRBC_ESP32_VL53L0X_H
+
+#include "mrubyc.h"
+
+void mrbc_mruby_esp32_vl53l0x_gem_init(struct VM*);
+
+#endif // MRBC_ESP32_VL53L0X_H
